Moved usePromise hook into helpers
