+++ src/common/helpers.ts
@@ -1,6 +1,6 @@
 import { AuthData } from "./services/discord";
 import { getBackendHost } from "./const";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 
 export const getStoreUrl = (id: string) => {
     return `https://www.microsoft.com/store/apps/${id}`;
@@ -57,6 +57,28 @@ export async function fetchBackend(route: string, method: ("GET" | "POST" | "PUT
         body: JSON.stringify(requestBody)
     });
 }
+
+
+export interface IUsePromiseState<T> {
+    results?: T
+    error?: Error
+    isLoading: boolean
+}
+
+export function usePromise<T>(promise: (Promise<T>)): IUsePromiseState<T> {
+    const [visualState, setVisualState] = useState<IUsePromiseState<T>>({ isLoading: true });
+
+    useEffect(() => {
+        promise.then(results => {
+            setVisualState(prevState => ({ ...prevState, isLoading: false, results }));
+        }).catch(error => {
+            setVisualState(prevState => ({ ...prevState, isLoading: false, error }));
+        });
+    }, [visualState]);
+
+    return visualState;
+};
+
 /**
  * @summary Get the first matching regex group, instead of an array with the full string and all matches
  * @param {string} toMatch
